proxy-server/services/requirements.txt
# DeepSeek-OCR Server Dependencies
# Python 3.8+ required

# Core dependencies
torch>=2.0.0
transformers>=4.35.0
pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Optional but recommended
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# For OCR service
requests>=2.31.0
python-dotenv>=1.0.0
